Android: Support SDL 2.26.4
